Don't chdir for test files
Fixes unused-result warning for chdir which is annoteded with
warn_unused_result in glibc.

Relative to cwd won't likely get you to those files in Samples, should
be relative to program directory but as we use cmake to run the tests
this becomes kinda moot. So just drop the call.

Signed-off-by: Ralph Sennhauser <ralph.sennhauser@gmail.com>