Move tests and test runner out of library
This means the runner is no longer part of the library. Users may still
grab it from the tests directory if they need but there are modern
alternatives out there specifically for the purpose of unit testing.

Also some test weren't run prior to this move, which might have been an
accident as they pass on the local machine.

Signed-off-by: Ralph Sennhauser <ralph.sennhauser@gmail.com>