Don't call xmlCleanupParser
Straight from the docs: WARNING: if your application is multithreaded or
has plugin support calling this may crash the application if another
thread or a plugin is still using libxml2. It's sometimes very hard to
guess if libxml2 is in use in the application, some libraries or plugins
may use it without notice. In case of doubt abstain from calling this
function or do it just before calling exit() to avoid leak reports from
valgrind !

https://gnome.pages.gitlab.gnome.org/libxml2/devhelp/libxml2-parser.html#xmlCleanupParser

Signed-off-by: Ralph Sennhauser <ralph.sennhauser@gmail.com>